バッチ設計_DBE517001_審査会資料一括作成（委員）

+++ DB/DBE/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbe/definition/processprm/shiryoshinsakai/IinTokkiJikouItiziHanteiProcessParameter.java
@@ -99,7 +99,7 @@ public class IinTokkiJikouItiziHanteiProcessParameter implements IBatchProcessPa
         if (RString.isNullOrEmpty(shuturyokuJun)) {
             isShuturyokuJun = true;
         }
-        return new IinTokkiJikouItiziHanteiMyBatisParameter(shinsakaiKaisaiNo, shuturyokuJun, bangoStart, bangoEnd,
+        return new IinTokkiJikouItiziHanteiMyBatisParameter(shinsakaiKaisaiNo, shuturyokuJun, bangoStart.intValue(), bangoEnd.intValue(),
                 isSakuseiJokenHani, isSakuseiJokenTuika, isShuturyokuJun);
     }
